wip: moved a bunch of classes around [macata #2]
