+++ src/main/java/com/dazednconfused/catalauncher/database/base/DAOException.java
@@ -1,4 +1,4 @@
-package com.dazednconfused.catalauncher.database;
+package com.dazednconfused.catalauncher.database.base;
 
 /**
  * Custom {@link Exception} wrapper for {@link Throwable}s at the {@link BaseDAO} layer.
