+++ src/main/java/com/dazednconfused/catalauncher/database/base/NotImplementedException.java
@@ -1,4 +1,4 @@
-package com.dazednconfused.catalauncher.database;
+package com.dazednconfused.catalauncher.database.base;
 
 /**
  * Represents unimplemented method invocations.
