+++ src/main/java/com/dazednconfused/catalauncher/database/mod/repository/ModH2RepositoryImpl.java
@@ -1,7 +1,7 @@
 package com.dazednconfused.catalauncher.database.mod.repository;
 
-import com.dazednconfused.catalauncher.database.DAOException;
-import com.dazednconfused.catalauncher.database.migration.MigrateableH2Database;
+import com.dazednconfused.catalauncher.database.base.DAOException;
+import com.dazednconfused.catalauncher.database.h2.migration.MigrateableH2Database;
 import com.dazednconfused.catalauncher.database.mod.dao.ModDAO;
 import com.dazednconfused.catalauncher.database.mod.dao.ModfileDAO;
 import com.dazednconfused.catalauncher.database.mod.entity.ModEntity;
